requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langgraph==0.1.19
langsmith==0.1.98
loguru==0.7.2
MarkupSafe==2.1.5
marshmallow==3.21.3
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.10.1
numpy==1.26.4
openai==1.39.0
openapi-schema-pydantic==1.2.4
orjson==3.10.6
packaging==24.1
psycopg2==2.9.9
pydantic==1.10.17
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
